update on qa permit calculation details

+++ apollo-store/src/main/kotlin/org/kebs/app/kotlin/apollo/store/model/qa/ManufacturePlantDetailsEntity.kt
@@ -33,7 +33,7 @@ class ManufacturePlantDetailsEntity : Serializable {
 
     @Column(name = "INSPECTION_FEE_STATUS")
     @Basic
-    var inspectionFeeStatus: Long? = null
+    var inspectionFeeStatus: Int? = null
 
     @Column(name = "PAID_DATE")
     @Basic
